TraceDag: split into interface and impl
This will make testing easier.
